Split rtl8150.c in two - rtl8150.h is the offspring.
skb_pool crap removed.

Async register write is a lot saner.  The register value used to be in a
statically allocated buffer so two concurrently running URBs would mess it's
content.  Not that this really matters as the last write would be the one that
counts, but anyway.

Reading and writing rtl8150's registers will print error when in debug mode.

The code is not yet tested on a real device, but at least compiles cleanly.